+++ package.json
@@ -39,6 +39,7 @@
     "react-scripts": "4.0.3",
     "react-transition-group": "^4.4.2",
     "resize-observer-polyfill": "^1.5.1",
+    "tplant": "^3.1.0",
     "typescript": "^4.3.5"
   },
   "scripts": {
