+++ src/components/DiagramEditor/ComponentsList/ComponentsList.module.scss
@@ -1,0 +1,8 @@
+.list {
+  position: absolute;
+  width: 100%;
+  max-width: 200px;
+  left: 30px;
+  top: 30px;
+  z-index: 1;
+}
