+++ src/components/DiagramEditor/Diagram/models/ClassPortModel.ts
@@ -1,0 +1,17 @@
+import {
+    LinkModel, PortModel, DefaultLinkModel, PortModelAlignment,
+} from '@projectstorm/react-diagrams';
+
+export class ClassPortModel extends PortModel {
+    constructor(alignment: PortModelAlignment) {
+        super({
+            type: 'class',
+            name: alignment,
+            alignment,
+        });
+    }
+
+    createLinkModel(): LinkModel {
+        return new DefaultLinkModel();
+    }
+}
