+++ src/components/DiagramEditor/Diagram/widgets/ClassNodeWidget/ClassNodeWidget.module.scss
