+++ src/components/DiagramEditor/Diagram/widgets/DiagramWidget/DiagramWidget.module.scss
@@ -1,0 +1,9 @@
+.canvas {
+  height: 100%;
+  display: flex;
+  width: 100%;
+  & > div {
+    width: 100%;
+  }
+}
+
