--- src/components/uml-editor/ItemsCreator/constants.ts
@@ -1,1 +1,0 @@
-export const NODE_CREATOR_EVENT_FORMAT = 'uml-diagram-node';
