--- src/components/uml-editor/ItemsCreator/styles/Item.module.scss
@@ -1,3 +1,0 @@
-.item {
-
-}
