--- src/components/uml-editor/ItemsCreator/styles/NodesCreator.module.scss
@@ -1,8 +1,0 @@
-.container {
-  position: absolute;
-  width: 100%;
-  max-width: 200px;
-  left: 30px;
-  top: 30px;
-  z-index: 1;
-}
