--- src/components/uml-editor/Widget/diagram/constants.ts
@@ -1,9 +1,0 @@
-export const NODE_TYPES = {
-    class: 'class',
-    interface: 'interface',
-};
-
-export enum PORT_TYPES {
-    method = 'method',
-    property = 'property'
-}
