--- src/components/uml-editor/Widget/diagram/methods/types.ts
@@ -1,4 +1,0 @@
-export interface Method {
-    name: string
-    key:string
-}
