--- src/components/uml-editor/Widget/diagram/nodes/class/ClassPortModel.ts
@@ -1,18 +1,0 @@
-import {
-    LinkModel, PortModel, DefaultLinkModel, PortModelAlignment,
-} from '@projectstorm/react-diagrams';
-import { NODE_TYPES } from '../../constants';
-
-export class ClassPortModel extends PortModel {
-    constructor(alignment: PortModelAlignment) {
-        super({
-            type: NODE_TYPES.class,
-            name: alignment,
-            alignment,
-        });
-    }
-
-    createLinkModel(): LinkModel {
-        return new DefaultLinkModel();
-    }
-}
