--- src/components/uml-editor/Widget/diagram/properties/types.ts
@@ -1,3 +1,0 @@
-export interface Property {
-    name: string
-}
