--- src/components/uml-editor/Widget/styles/Widget.module.scss
@@ -1,8 +1,0 @@
-.widget {
-  height: 100%;
-  display: flex;
-  width: 100%;
-  & > div {
-    width: 100%;
-  }
-}
