+++ src/locales/en/messages.js
@@ -1,1 +1,1 @@
-/*eslint-disable*/module.exports={messages:{"ADD_METHOD":"Add method","ADD_PROPERTY":"add property","NEW_METHOD":"New method","NEW_PROPERTY":"New property","NODE_TYPE_CLASS":"Class","NODE_TYPE_INTERFACE":"Interface"}};
+/*eslint-disable*/module.exports={messages:{"ADD_METHOD":"Add method","ADD_PROPERTY":"add property","COMPONENTS_NAMES_CLASS":"Class","COMPONENTS_NAMES_ENUM":"Enum","COMPONENTS_NAMES_INTERFACE":"Interface","COMPONENTS_NAMES_NAMESPACE":"Namespace","NEW_METHOD":"New method","NEW_PROPERTY":"New property","NODE_TYPE_CLASS":"Class","NODE_TYPE_INTERFACE":"Interface"}};
