+++ src/locales/lang-constants.ts
@@ -1,0 +1,8 @@
+import { t } from '@lingui/macro';
+
+export const COMPONENTS_NAMES: { [key: number]: string } = {
+    0: t`COMPONENTS_NAMES_CLASS`,
+    1: t`COMPONENTS_NAMES_ENUM`,
+    4: t`COMPONENTS_NAMES_INTERFACE`,
+    6: t`COMPONENTS_NAMES_NAMESPACE`,
+};
