+++ src/locales/ru/messages.po
@@ -13,15 +13,33 @@ msgstr ""
 "Language-Team: \n"
 "Plural-Forms: \n"
 
-#: src/components/uml-editor/Widget/diagram/nodes/class/ClassNodeWidget.tsx:56
+#: src/components/DiagramEditor/Diagram/widgets/ClassNodeWidget/index.tsx:58
+#: src/components/uml-editor/Widget/diagram/nodes/class/ClassNodeWidget.tsx:59
 msgid "ADD_METHOD"
 msgstr "Добавить метод"
 
-#: src/components/uml-editor/Widget/diagram/nodes/class/ClassNodeWidget.tsx:57
+#: src/components/DiagramEditor/Diagram/widgets/ClassNodeWidget/index.tsx:59
+#: src/components/uml-editor/Widget/diagram/nodes/class/ClassNodeWidget.tsx:60
 msgid "ADD_PROPERTY"
 msgstr "Добавить свойство"
 
-#: src/components/uml-editor/Widget/diagram/methods/MethodModel.ts:13
+#: src/locales/lang-constants.ts:4
+msgid "COMPONENTS_NAMES_CLASS"
+msgstr "класс"
+
+#: src/locales/lang-constants.ts:5
+msgid "COMPONENTS_NAMES_ENUM"
+msgstr "Перечисление"
+
+#: src/locales/lang-constants.ts:6
+msgid "COMPONENTS_NAMES_INTERFACE"
+msgstr "Интерфейс"
+
+#: src/locales/lang-constants.ts:7
+msgid "COMPONENTS_NAMES_NAMESPACE"
+msgstr "Пространство имен"
+
+#: src/components/uml-editor/Widget/diagram/methods/MethodModel.ts:16
 msgid "NEW_METHOD"
 msgstr "Новый метод"
 
@@ -29,10 +47,10 @@ msgstr "Новый метод"
 msgid "NEW_PROPERTY"
 msgstr "Новое свойство"
 
-#: src/components/uml-editor/items/ItemClass.ts:11
+#: src/components/uml-editor/items/ItemClass.ts:17
 msgid "NODE_TYPE_CLASS"
 msgstr "Класс"
 
-#: src/components/uml-editor/items/ItemInterface.ts:11
+#: src/components/uml-editor/items/ItemInterface.ts:17
 msgid "NODE_TYPE_INTERFACE"
 msgstr "Интерфейс"
