+++ src/models/Diagram.ts
@@ -1,0 +1,8 @@
+import { DiagramEngine } from '@projectstorm/react-diagrams';
+import { Point } from '@projectstorm/geometry';
+import { ComponentKind } from 'tplant/dist/Models/ComponentKind';
+
+export default interface Diagram {
+    engine(): DiagramEngine
+    addNode(item: ComponentKind, point?: Point): void
+}
