+++ src/pages/main/index.tsx
@@ -1,11 +1,11 @@
-import UmlEditor from '../../components/uml-editor';
+import DiagramEditor from '../../components/DiagramEditor';
 
 /**
  * Component serving the main page of the application.
  */
 function PageMain() {
     return (
-        <UmlEditor />
+        <DiagramEditor />
     );
 }
